move AbstractAuditTrailManager
