+++ settings.gradle
@@ -67,7 +67,6 @@ include "support:cas-server-support-person-directory"
 include "support:cas-server-support-bootadmin-client"
 include "support:cas-server-support-jdbc-drivers"
 include "support:cas-server-support-metrics"
-include "support:cas-server-support-audit-core"
 include "support:cas-server-support-audit-mongo"
 include "support:cas-server-support-audit-jdbc"
 include "support:cas-server-support-audit-couchbase"
