--- support/cas-server-support-audit-core/build.gradle
@@ -1,5 +1,0 @@
-description = "Apereo CAS Couchbase Audit Support"
-dependencies {
-    api project(":api:cas-server-core-api")
-    api project(":api:cas-server-core-api-audit")
-}
