+++ support/cas-server-support-audit-couchbase/build.gradle
@@ -11,7 +11,6 @@ dependencies {
     implementation project(":core:cas-server-core-util-api")
     implementation project(":core:cas-server-core-configuration-api")
     implementation project(":core:cas-server-core-services-registry")
-    implementation project(":support:cas-server-support-audit-core")
     implementation project(":support:cas-server-support-couchbase-core")
     
     testImplementation project(":support:cas-server-support-person-directory")
