+++ support/cas-server-support-audit-jdbc/build.gradle
@@ -5,7 +5,6 @@ dependencies {
 
     implementation project(":core:cas-server-core-audit-api")
     implementation project(":core:cas-server-core-util-api")
-    implementation project(":support:cas-server-support-audit-core")
 
     testImplementation project(":support:cas-server-support-person-directory")
     testImplementation project(":core:cas-server-core-services")
