+++ support/cas-server-support-audit-mongo/build.gradle
@@ -7,7 +7,6 @@ dependencies {
     
     implementation project(":core:cas-server-core-util-api")
     implementation project(":core:cas-server-core-audit-api")
-    implementation project(":support:cas-server-support-audit-core")
     implementation project(":support:cas-server-support-mongo-core")
     
     testImplementation project(":support:cas-server-support-person-directory")
