+++ support/cas-server-support-audit-mongo/src/main/java/org/apereo/cas/audit/MongoDbAuditTrailManager.java
@@ -1,5 +1,6 @@
 package org.apereo.cas.audit;
 
+import org.apereo.cas.audit.spi.AbstractAuditTrailManager;
 import org.apereo.cas.util.DateTimeUtils;
 
 import lombok.RequiredArgsConstructor;
